requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.0
requests==2.31.0
